requirements.txt
alembic==1.16.4
anyio==4.9.0
certifi==2025.6.15
emoji==2.14.1
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.25.2
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.2
python-dotenv==1.1.1
python-telegram-bot==20.7
pytz==2025.2
sniffio==1.3.1
SQLAlchemy==2.0.41
typing_extensions==4.14.0
watchdog==6.0.0
